issue_2510 [Improvement]: add some flink runtime info for optimizer
